+++ R/utils-mocks.R
@@ -6,5 +6,5 @@ file.remove_ <- function(...) file.remove(...)
 file.exists_ <- function(...) file.exists(...)
 is_interactive <- function(...) interactive()
 list.files_ <- function(...) list.files(...)
-
 load_all <- function(...) devtools::load_all(...)
+menu <- function(...) utils::menu(...)