+++ R/utils.R
@@ -1,1 +1,11 @@
-placeholder <- function() message("Hello")
+devtools_load <- function(load = TRUE) {
+    checkmate::assert_flag(load)
+
+    if (isTRUE(load)) {
+        load_all()
+
+        return(devtools_load(FALSE))
+    } else {
+        invisible(NULL)
+    }
+}